Fix tournament match creation and match_consumer socket closing
