+++ backend/pong/templates/pong.html
@@ -3,16 +3,10 @@
 <head>
     <title>Pong Game</title>
     <style>
-        #gameCanvas {
-            background-color: #000;
-            display: none; /* Hide the canvas by default */
-            margin: 0 auto;
-        }
     </style>
     {% load static %}
 </head>
 <body>
-    <canvas id="gameCanvas" width="1600" height="900"></canvas>
     <button id="join-queue-button">Join Queue</button>
     <button id="cancel-queue-button" style="display: none;">Cancel Queue</button>
     <button id="reconnectButton" style="display: none;">Reconnect</button>
