+++ backend/pong/templates/tournament.html
@@ -21,6 +21,9 @@
     <h2>Open Tournaments</h2>
     <div id="tournament-list"></div>
 
+    <!-- Load match.js as a module -->
+    <script type="module" src="{% static 'match.js' %}"></script>
+
     <!-- Load matchmaking.js as a module -->
     <script type="module" src="{% static 'matchmaking.js' %}"></script>
